finished app volume 1 (base)

+++ web/public/php/remove-all.php
@@ -2,7 +2,11 @@
 require_once '../../app/vendor/autoload.php';
 use \App\Datatype\DB;
 
-$pdo = (new DB)->get_pdo();
-$sql = "DELETE FROM names";
-$query = $pdo->query($sql);
-header('location: /');
+$is_confirmed = htmlspecialchars($_GET['is_confirmed']);
+if ($is_confirmed === 'yes') {
+    $pdo = (new DB)->get_pdo();
+    $sql = "DELETE FROM names";
+    $query = $pdo->query($sql);
+} else {
+    echo 'Ошибка сервера. Повторите позже.';
+}
